--- MiniProject/.idea/libraries/Maven__com_netflix_netflix_commons_netflix_infix_0_3_0.xml
@@ -1,13 +1,0 @@
-<component name="libraryTable">
-  <library name="Maven: com.netflix.netflix-commons:netflix-infix:0.3.0">
-    <CLASSES>
-      <root url="jar://$MAVEN_REPOSITORY$/com/netflix/netflix-commons/netflix-infix/0.3.0/netflix-infix-0.3.0.jar!/" />
-    </CLASSES>
-    <JAVADOC>
-      <root url="jar://$MAVEN_REPOSITORY$/com/netflix/netflix-commons/netflix-infix/0.3.0/netflix-infix-0.3.0-javadoc.jar!/" />
-    </JAVADOC>
-    <SOURCES>
-      <root url="jar://$MAVEN_REPOSITORY$/com/netflix/netflix-commons/netflix-infix/0.3.0/netflix-infix-0.3.0-sources.jar!/" />
-    </SOURCES>
-  </library>
-</component>
