--- MiniProject/.idea/libraries/Maven__commons_jxpath_commons_jxpath_1_3.xml
@@ -1,13 +1,0 @@
-<component name="libraryTable">
-  <library name="Maven: commons-jxpath:commons-jxpath:1.3">
-    <CLASSES>
-      <root url="jar://$MAVEN_REPOSITORY$/commons-jxpath/commons-jxpath/1.3/commons-jxpath-1.3.jar!/" />
-    </CLASSES>
-    <JAVADOC>
-      <root url="jar://$MAVEN_REPOSITORY$/commons-jxpath/commons-jxpath/1.3/commons-jxpath-1.3-javadoc.jar!/" />
-    </JAVADOC>
-    <SOURCES>
-      <root url="jar://$MAVEN_REPOSITORY$/commons-jxpath/commons-jxpath/1.3/commons-jxpath-1.3-sources.jar!/" />
-    </SOURCES>
-  </library>
-</component>
