--- MiniProject/.idea/libraries/Maven__joda_time_joda_time_2_3.xml
@@ -1,13 +1,0 @@
-<component name="libraryTable">
-  <library name="Maven: joda-time:joda-time:2.3">
-    <CLASSES>
-      <root url="jar://$MAVEN_REPOSITORY$/joda-time/joda-time/2.3/joda-time-2.3.jar!/" />
-    </CLASSES>
-    <JAVADOC>
-      <root url="jar://$MAVEN_REPOSITORY$/joda-time/joda-time/2.3/joda-time-2.3-javadoc.jar!/" />
-    </JAVADOC>
-    <SOURCES>
-      <root url="jar://$MAVEN_REPOSITORY$/joda-time/joda-time/2.3/joda-time-2.3-sources.jar!/" />
-    </SOURCES>
-  </library>
-</component>
