--- MiniProject/.idea/libraries/Maven__org_antlr_antlr_runtime_3_4.xml
@@ -1,13 +1,0 @@
-<component name="libraryTable">
-  <library name="Maven: org.antlr:antlr-runtime:3.4">
-    <CLASSES>
-      <root url="jar://$MAVEN_REPOSITORY$/org/antlr/antlr-runtime/3.4/antlr-runtime-3.4.jar!/" />
-    </CLASSES>
-    <JAVADOC>
-      <root url="jar://$MAVEN_REPOSITORY$/org/antlr/antlr-runtime/3.4/antlr-runtime-3.4-javadoc.jar!/" />
-    </JAVADOC>
-    <SOURCES>
-      <root url="jar://$MAVEN_REPOSITORY$/org/antlr/antlr-runtime/3.4/antlr-runtime-3.4-sources.jar!/" />
-    </SOURCES>
-  </library>
-</component>
