--- MiniProject/.idea/libraries/Maven__org_apache_commons_commons_math_2_2.xml
@@ -1,13 +1,0 @@
-<component name="libraryTable">
-  <library name="Maven: org.apache.commons:commons-math:2.2">
-    <CLASSES>
-      <root url="jar://$MAVEN_REPOSITORY$/org/apache/commons/commons-math/2.2/commons-math-2.2.jar!/" />
-    </CLASSES>
-    <JAVADOC>
-      <root url="jar://$MAVEN_REPOSITORY$/org/apache/commons/commons-math/2.2/commons-math-2.2-javadoc.jar!/" />
-    </JAVADOC>
-    <SOURCES>
-      <root url="jar://$MAVEN_REPOSITORY$/org/apache/commons/commons-math/2.2/commons-math-2.2-sources.jar!/" />
-    </SOURCES>
-  </library>
-</component>
