+++ MiniProject/CoreModule/src/main/java/com/mini_project/CoreModule/CoremoduleApplication.java
