+++ MiniProject/CoreModule/src/main/java/com/mini_project/CoreModule/entity/EmbargoEntity.java
