+++ MiniProject/ExportModule/src/main/java/com/haonguyen/ExportService/repository/ICommodityRepository.java
@@ -1,0 +1,2 @@
+package com.haonguyen.ExportService.repository;public interface ICommodityRepository {
+}
