+++ MiniProject/ExportModule/src/main/java/com/haonguyen/ExportService/repository/ISectorsEntity.java
@@ -1,0 +1,2 @@
+package com.haonguyen.ExportService.repository;public class ISectorsEntity {
+}
