+++ MiniProject/ExportModule/src/main/java/com/haonguyen/ExportService/service/CommodityService.java
@@ -1,0 +1,2 @@
+package com.haonguyen.ExportService.service;public class CommodityService {
+}
