+++ MiniProject/ExportModule/src/main/java/com/haonguyen/ExportService/service/SectorsService.java
@@ -1,0 +1,2 @@
+package com.haonguyen.ExportService.service;public class SectorsService {
+}
